feat: 添加API响应状态检查功能
- 在types.go中新增SuccessCode常量用于标识成功响应状态码200
- 在多个交易相关的WebSocket API方法中添加状态码验证逻辑
- 当API响应状态不为SuccessCode时，返回对应的API错误信息
- 统一处理订单放置、修改、取消、状态查询等操作的错误响应
- 改进算法订单相关服务的错误处理机制

+++ websocket_api.go
@@ -53,7 +53,7 @@ type WsAPIRateLimit struct {
 }
 
 type WsAPIErrorResponse struct {
-	Code    int    `json:"code"`
+	Code    int64  `json:"code"`
 	ID      string `json:"id"`
 	Message string `json:"msg"`
 }
